Move constants into shared.
