+++ lsif/src/shared/constants.ts
