+++ lsif/src/shared/xrepo/xrepo.ts
@@ -13,7 +13,7 @@ import { TableInserter } from '../database/inserter'
 import { addrFor, getCommitsNear, gitserverExecLines } from './commits'
 import { TracingContext, logAndTraceCall } from '../tracing'
 import { dbFilename, tryDeleteFile } from '../util'
-import { MAX_TRAVERSAL_LIMIT, ADVISORY_LOCK_ID_SALT, MAX_CONCURRENT_GITSERVER_REQUESTS } from '../../constants'
+import { MAX_TRAVERSAL_LIMIT, ADVISORY_LOCK_ID_SALT, MAX_CONCURRENT_GITSERVER_REQUESTS } from '../constants'
 import { chunk } from 'lodash'
 
 /**
