+++ lsif/src/tests/reference-pagination-monorepo.test.ts
@@ -1,7 +1,7 @@
 import { BackendTestContext, filterNodeModules } from './util'
 import { createCommit, createLocation } from '../test-utils'
 import { ReferencePaginationContext } from '../server/backend/backend'
-import { MAX_TRAVERSAL_LIMIT } from '../constants'
+import { MAX_TRAVERSAL_LIMIT } from '../shared/constants'
 import { lsp } from 'lsif-protocol'
 
 describe('Backend', () => {
