+++ lsif/src/worker/processors/clean-failed-jobs.ts
@@ -1,4 +1,4 @@
-import * as constants from '../../constants'
+import * as constants from '../../shared/constants'
 import * as fs from 'mz/fs'
 import * as path from 'path'
 import { FAILED_JOB_MAX_AGE, STORAGE_ROOT } from '../settings'
